Fixes for running as a normal postgres user.
The pg_largeobject table can only be queried as a superuser. This fixes
things by using the pg_largeobject_metadata table instead.

The permission issues also mean we can't really implement an efficient
server-side copy function, so we'll revert back to the default copy
functionality (so we'll fetch the data down, and then upload it back).